+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add default octicons for custom highlights (thanks to [Andrew Hlynskyi](https://github.com/ahlinc))
 - Always show total in the activity bar regardless of other settings
+- Add support for hiding counts from the activity bar badge
 
 ## v0.0.220 - 2022-11-07